pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gradual-mining"
version = "0.1.0"
description = "Unified Python package for gradual pattern mining"
readme = "README.md"
authors = [
    {name = ""}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["data mining", "pattern mining", "gradual patterns", "knowledge discovery"]

dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "mypy>=0.950",
    "flake8>=4.0",
]
seasonal = [
    "jpype1>=1.4.0",  # For SPMF integration
]
fuzzy = [
    "scikit-fuzzy>=0.4.0",  # For temporal patterns
]
all = [
    "gradual-mining[dev,seasonal,fuzzy]",
]

[project.scripts]
gradual-mine = "gradual_mining.cli:main"

[project.urls]
Homepage = ""
Documentation = ""
Repository = ""

[tool.setuptools]
packages = ["gradual_mining"]

[tool.setuptools.package-data]
gradual_mining = ["py.typed"]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--verbose --cov=gradual_mining --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
